Add UInt24.

+++ src/MrKWatkins.BinaryPrimitives.Tests/ByteIReadOnlyListExtensionsTests.cs
@@ -64,8 +64,8 @@ public sealed class ByteIReadOnlyListExtensionsTests
     {
         IReadOnlyList<byte> bytes = [0x78, 0x56, 0x34, 0x12];
 
-        bytes.GetUInt24(0).Should().Equal(0x345678);
-        bytes.GetUInt24(1).Should().Equal(0x123456);
+        ((int)bytes.GetUInt24(0)).Should().Equal(0x345678);
+        ((int)bytes.GetUInt24(1)).Should().Equal(0x123456);
     }
 
 
@@ -74,10 +74,10 @@ public sealed class ByteIReadOnlyListExtensionsTests
     {
         IReadOnlyList<byte> bytes = [0x78, 0x56, 0x34, 0x12];
 
-        bytes.GetUInt24(0, Endian.Little).Should().Equal(0x345678);
-        bytes.GetUInt24(1, Endian.Little).Should().Equal(0x123456);
-        bytes.GetUInt24(0, Endian.Big).Should().Equal(0x785634);
-        bytes.GetUInt24(1, Endian.Big).Should().Equal(0x563412);
+        ((int)bytes.GetUInt24(0, Endian.Little)).Should().Equal(0x345678);
+        ((int)bytes.GetUInt24(1, Endian.Little)).Should().Equal(0x123456);
+        ((int)bytes.GetUInt24(0, Endian.Big)).Should().Equal(0x785634);
+        ((int)bytes.GetUInt24(1, Endian.Big)).Should().Equal(0x563412);
     }
